pdf/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pandas==2.2.2
PyMuPDF==1.24.4
openpyxl==3.1.3
google-auth==2.29.0
requests==2.32.3